add component kafka listener

+++ TFM/02_DatabaseAsAServiceInterface/v2/externalConsumer/src/main/java/es/codeurjc/externalConsumer/repositories/OrderRepository.java
@@ -7,6 +7,6 @@ import java.util.Optional;
 
 public interface OrderRepository extends JpaRepository<Order, Long> {
 
-    Optional<Order> findById(Integer id);
+    Optional<Order> findById(Long id);
 
 }
